+++ app/components/render-tooltips.js
@@ -1,0 +1,1 @@
+export { default } from '@zestia/ember-async-tooltips/components/render-tooltips';